==> Methods/SpatialMap/MERFISH/STEP3_Prediction.R <==

rm(list = ls())
source("funcs.R")
library(MASS)
library(Matrix)
para="car2"
######################## STEP 1: load information ######################## 
ST=readRDS("ST.rds")
X=readRDS("X.rds")
total=rowSums(ST)
n=nrow(ST)
sc=readRDS("scFull.rds")
K=read.table(paste0("K_car0.2.txt"))

filelist=dir("./Estimation")
file=filelist[grep(para, filelist)]
Estimation = data.frame()
for (i in 1:length(file)) {
  est=readRDS(paste0("./Estimation/",file[i]))
  Estimation=rbind(Estimation, est)
  cat(paste0("----",i,"----"),'\n')
}
## Remove NaN
index2=which(is.nan(Estimation$h2))
if(length(index2)==0){cat("No need to remove NaN")}else{Estimation=Estimation[-index2,]}
## Select by p-value
Estimation=Estimation[Estimation$pvalue<0.05,]

########################  STEP 2: Prepare reference ######################## 
TestNameList=read.csv("ReferGene.csv")
TestNameList=TestNameList$x
TestSC=sc[TestNameList,]
ReferNameList=setdiff(intersect(rownames(Estimation),rownames(sc)),TestNameList)
ReferSC=sc[ReferNameList,]
rm(sc)

########################  STEP 3: Prediction ######################## 
Prediction=data.frame()
for(l in 1:length(TestNameList)){
  Gene=TestNameList[l]
  cat(paste0("----- Predicting: ", Gene," -------"),'\n')
  Correlation=cor(TestSC[Gene,],t(ReferSC))
  ReferName=colnames(Correlation)[which.max(Correlation)]
  
  estimates=Estimation[ReferName,]
  alpha=estimates$alpha
  beta=estimates$beta
  h2=estimates$h2
  sigma2=estimates$sigma2
  SigmaG=h2*sigma2*K
  SigmaE=diag(rep((1-h2)*sigma2,n))
  
  g=mvrnorm(n=100, mu=rep(0,n),SigmaG)
  g=apply(g, 2, mean)
  e=mvrnorm(n=100, mu=rep(0,n),SigmaE)
  e=apply(e, 2, mean)
  
  lambda=exp(alpha+beta*X[,Gene]+g+e)
  lambda=lambda*10
  p=apply(matrix(lambda*total),1, function(x){round(mean(rpois(100,x)))})
  Prediction=rbind(Prediction,p)
}

rownames(Prediction)=TestNameList
colnames(Prediction)=paste0("sites",1:n)
####################### STEP 4: Record ######################
ofile=paste0(para,"_ReferGene.csv")
write.csv(Prediction, file=ofile)
